# api/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.1",
    "title": "Quiz API",
    "description": "Manage Quizzes"
  },
  "servers": [
    {
      "url": "http://localhost:4022/api"
    }
  ],
...
  "securityDefinitions": {},
  "paths": {
...
    "/quizzes/{quizId}/questions/{questionId}": {
      "get": {
        "tags": [
          "Question"
        ],
        "description": "Get one question",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "404": {
            "description": "Question not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "read:questions"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Question"
        ],
        "description": "Delete a question",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question deleted successfully"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "delete:questions"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Question"
        ],
        "description": "Update a question",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "update:questions"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestion"
              }
            }
          }
        }
      }
    },
    "/quizzes/{quizId}/questions": {
      "post": {
        "tags": [
          "Question"
        ],
        "description": "Create a question",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "create:questions"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddQuestion"
              }
            }
          }
        }
      }
    },
    "/questions": {
      "get": {
        "tags": [
          "Question"
        ],
        "description": "Get all questions",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "How many elements to skip. Used for paging.",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "in": "query",
            "description": "How many elements to take. Used for paging.",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort by field",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "in": "query",
            "description": "Search by name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quizzes",
            "in": "query",
            "description": "Search by list of quizzes. Example: 1,2,3",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topics",
            "in": "query",
            "description": "Search by list of topics. Example: 1,2,3",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questions fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionSearchResult"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionSearchResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "read:questions"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "auth0": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://dev-ki47hyci.us.auth0.com/authorize",
            "scopes": [
              "read:quizzes",
              "create:quizzes"
            ]
          }
        }
      }
    },
    "schemas": {
...
      "AddQuestion": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "what?"
          },
          "choice0": {
            "type": "string",
            "example": "one"
          },
          "choice1": {
            "type": "string",
            "example": "two"
          },
          "choice2": {
            "type": "string",
            "example": "three"
          },
          "choice3": {
            "type": "string",
            "example": "four"
          },
          "correct0": {
            "type": "number",
            "example": 1
          },
          "correct1": {
            "type": "number",
            "example": 0
          },
          "correct2": {
            "type": "number",
            "example": 0
          },
          "correct3": {
            "type": "number",
            "example": 0
          }
        },
        "xml": {
          "name": "AddQuestion"
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "text": {
            "type": "string",
            "example": "what?"
          },
          "choice0": {
            "type": "string",
            "example": "one"
          },
          "choice1": {
            "type": "string",
            "example": "two"
          },
          "choice2": {
            "type": "string",
            "example": "three"
          },
          "choice3": {
            "type": "string",
            "example": "four"
          },
          "correct0": {
            "type": "number",
            "example": 1
          },
          "correct1": {
            "type": "number",
            "example": 0
          },
          "correct2": {
            "type": "number",
            "example": 0
          },
          "correct3": {
            "type": "number",
            "example": 0
          },
          "createdBy": {
            "type": "number",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "example": "2020-03-31T00:00:00.000Z"
          }
        },
        "xml": {
          "name": "Question"
        }
      },
      "UpdateQuestion": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "what?"
          },
          "choice0": {
            "type": "string",
            "example": "one"
          },
          "choice1": {
            "type": "string",
            "example": "two"
          },
          "choice2": {
            "type": "string",
            "example": "three"
          },
          "choice3": {
            "type": "string",
            "example": "four"
          },
          "correct0": {
            "type": "number",
            "example": 1
          },
          "correct1": {
            "type": "number",
            "example": 0
          },
          "correct2": {
            "type": "number",
            "example": 0
          },
          "correct3": {
            "type": "number",
            "example": 0
          }
        },
        "xml": {
          "name": "UpdateQuestion"
        }
      },
...
      "QuestionArray": {
        "type": "array",
        "items": {
          "xml": {
            "name": "question"
          },
          "$ref": "#/components/schemas/Question"
        },
        "xml": {
          "name": "QuestionArray"
        }
      },
      "QuestionSearchResult": {
        "type": "object",
        "properties": {
          "search": {
            "type": "string",
            "example": "query string"
          },
          "topics": {
            "type": "string",
            "example": "1,2,3"
          },
          "quizzes": {
            "type": "string",
            "example": "1,2,3"
          },
          "sort": {
            "type": "string",
            "example": "name"
          },
          "order": {
            "type": "string",
            "example": "asc"
          },
          "skip": {
            "type": "number",
            "example": 0
          },
          "take": {
            "type": "number",
            "example": 10
          },
          "total": {
            "type": "number",
            "example": 100
          },
          "items": {
            "type": "array",
            "items": {
              "xml": {
                "name": "questionarray"
              },
              "$ref": "#/components/schemas/QuestionArray"
            }
          }
        },
        "xml": {
          "name": "QuestionSearchResult"
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": [],
      "auth0": []
    }
  ]
}